+++ packages/components/src/spectrum/buttons.ts
@@ -1,8 +1,10 @@
 export {
   ActionButton,
   type SpectrumActionButtonProps as ActionButtonProps,
-  ActionGroup,
-  type SpectrumActionGroupProps as ActionGroupProps,
+  // Button - we want to use our own `Button` component instead of Spectrum's
+  ButtonGroup,
+  type SpectrumButtonGroupProps as ButtonGroupProps,
+  // FileTrigger - we aren't planning to support this component
   LogicButton,
   type SpectrumLogicButtonProps as LogicButtonProps,
   ToggleButton,