+++ packages/components/src/theme/SpectrumThemeProvider.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, useState } from 'react';
 import { Provider } from '@adobe/react-spectrum';
 import type { Theme } from '@react-types/provider';
-import shortid from 'shortid';
+import { nanoid } from 'nanoid';
 import { themeDHDefault } from '../spectrum/utils';
 
 export interface SpectrumThemeProviderProps {
@@ -27,7 +27,7 @@ export function SpectrumThemeProvider({
   // a unique ID is used per provider to force it to render the theme wrapper element inside portals
   // based on https://github.com/adobe/react-spectrum/issues/1697#issuecomment-999827266
   // won't be needed if https://github.com/adobe/react-spectrum/pull/2669 is merged
-  const [id] = useState(isPortal ? shortid() : null);
+  const [id] = useState(isPortal ? nanoid() : null);
 
   return (
     <Provider